# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Prime Max & Elite Prefab ERP CRM API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /leads:
    get:
      operationId: listLeads
      tags: [leads]
      summary: List leads
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: companyId
          in: query
          schema:
            type: integer
        - name: assignedTo
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: leadScore
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of leads
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Lead"
    post:
      operationId: createLead
      tags: [leads]
      summary: Create lead
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "201":
          description: Lead created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
  /leads/{id}:
    get:
      operationId: getLead
      tags: [leads]
      summary: Get lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    put:
      operationId: updateLead
      tags: [leads]
      summary: Update lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLeadBody"
      responses:
        "200":
          description: Lead updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Lead"
    delete:
      operationId: deleteLead
      tags: [leads]
      summary: Delete lead
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Lead deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /leads/pipeline:
    get:
      operationId: getLeadsPipeline
      tags: [leads]
      summary: Get leads grouped by stage for kanban
      parameters:
        - name: companyId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Leads by stage
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LeadsPipeline"

  # CONTACTS
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications for current user
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: List of notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{id}/read:
    post:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Notification marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
...
  /audit-logs:
    get:
      operationId: listAuditLogs
      tags: [audit-logs]
      summary: List audit logs
      parameters:
        - name: userId
          in: query
          schema:
            type: integer
        - name: action
          in: query
          schema:
            type: string
        - name: entity
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: List of audit logs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AuditLog"

  # DASHBOARD
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required: [success]
...
    Lead:
      type: object
      properties:
        id:
          type: integer
        leadNumber:
          type: string
        leadName:
          type: string
        companyName:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        location:
          type: string
        source:
          type: string
        requirementType:
          type: string
        quantity:
          type: number
        budget:
          type: number
        status:
          type: string
        assignedToId:
          type: integer
        assignedToName:
          type: string
        notes:
          type: string
        nextFollowUp:
          type: string
        leadScore:
          type: string
        companyId:
          type: integer
        companyRef:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required: [id, leadNumber, leadName, status, leadScore, createdAt]
    CreateLeadBody:
      type: object
      properties:
        leadName:
          type: string
        companyName:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        whatsapp:
          type: string
        email:
          type: string
        location:
          type: string
        source:
          type: string
        requirementType:
          type: string
        quantity:
          type: number
        budget:
          type: number
        status:
          type: string
        assignedToId:
          type: integer
        notes:
          type: string
        nextFollowUp:
          type: string
        leadScore:
          type: string
        companyId:
          type: integer
      required: [leadName, status]
    LeadsPipeline:
      type: object
      properties:
        new:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        contacted:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        qualified:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        siteVisit:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        quotationRequired:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        quotationSent:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        negotiation:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        won:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
        lost:
          type: array
          items:
            $ref: "#/components/schemas/Lead"
      required: [new, contacted, qualified, siteVisit, quotationRequired, quotationSent, negotiation, won, lost]
...
    Notification:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        message:
          type: string
        type:
          type: string
        isRead:
          type: boolean
        userId:
          type: integer
        entityType:
          type: string
        entityId:
          type: integer
        createdAt:
          type: string
      required: [id, title, message, type, isRead, userId, createdAt]
    AuditLog:
      type: object
      properties:
        id:
          type: integer
        userId:
          type: integer
        userName:
          type: string
        action:
          type: string
        entity:
          type: string
        entityId:
          type: integer
        details:
          type: string
        ipAddress:
          type: string
        createdAt:
          type: string
      required: [id, action, entity, createdAt]
...